Add workflow for UPI vSphere platform none
This commit adds a command script and test step
upi-conf-vsphere-platform-none that modify the
install-config file to use platform:none as the
platform type.
This commit also adds a new chain and workflow for
upi-conf-vsphere-platform-none-ovn reusing some
steps used in upi-vsphere workflow in addition to the
steps required for setting up hybrid OVN networking.